chore: rename karma tests to *.browser.spec.ts
