+++ packages/fetch-http-handler/karma.conf.js
@@ -2,7 +2,7 @@ process.env.CHROME_BIN = require("puppeteer").executablePath();
 module.exports = function (config) {
   config.set({
     frameworks: ["jasmine", "karma-typescript"],
-    files: ["src/stream-collector.ts", "src/stream-collector.spec.ts"],
+    files: ["src/stream-collector.ts", "src/stream-collector.browser.spec.ts"],
     exclude: ["**/*.d.ts"],
     preprocessors: {
       "**/*.ts": "karma-typescript"
